backend/requirements.txt
fastapi>=0.116.0
uvicorn>=0.35.0
python-dotenv>=1.0.1
pydantic>=2.11.7
pydantic-settings>=2.10.1
python-multipart>=0.0.20
openai>=1.101.0
google-generativeai>=0.8.5
httpx>=0.26.0
